Add new field to Vacancy
A salary can be described now using min and max values, currency,
and also units (e.g. 'hour', 'month', 'year', or 'project').

Employment type (e.g. 'full-time' or 'contract') can be specified
for each vacancy using separate field for that.

And the last, but not least, each vacancy can now be marked as
"remote position". That is going to be used for data analysis in
the future.